MAINT: `realtransforms` move docstrings to new files and rename `np` to `uarray`
